Set default value for `app.initApp()` when not defined
When loading an app, before calling `app.render()` it is first
initialized by `app.initApp()` to set the `apiUrl`, `isMasqueradable`,
`accountId` & `userId`.
That method will be called for every app (except `auth` & `core`) but
we want to allow developers to execute code before if needed.
Because of that, during the monsterization process, we check if a custom
`app.initApp` exists and use it as a wrapper for our mandatory code.